requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-telegram-bot==20.8
python-dotenv==1.0.1
apscheduler==3.10.4
fake-useragent==1.5.1
pytz==2023.3.post1
